+++ src/core/codecs/lucene90/fields_index.rs
@@ -23,15 +23,15 @@ pub(crate) trait FieldsIndex: TryClone {
     fn get_block_id(&mut self, doc_id: i32) -> Result<i64>;
 
     /// Get the start pointer of the block with the given ID.
-    fn get_block_start_pointer(&mut self, block_id: i64) -> Result<i64>;
+    fn get_block_start_pointer(&mut self, block_id: i64) -> Result<usize>;
 
     /// Get the number of bytes of the block with the given ID.
-    fn get_block_length(&mut self, block_id: i64) -> Result<i64>;
+    fn get_block_length(&mut self, block_id: i64) -> Result<usize>;
 
     /// Get the start pointer of the block that contains the given docID.
     /// This is a final method in the original struct, so it's implemented
     /// directly here.
-    fn get_start_pointer(&mut self, doc_id: i32) -> Result<i64> {
+    fn get_start_pointer(&mut self, doc_id: i32) -> Result<usize> {
         let block_id = self.get_block_id(doc_id)?;
         self.get_block_start_pointer(block_id)
     }
@@ -74,19 +74,19 @@ where
         }
     }
 
-    fn get_block_start_pointer(&mut self, block_id: i64) -> Result<i64> {
+    fn get_block_start_pointer(&mut self, block_id: i64) -> Result<usize> {
         match self {
             FieldsIndexEnum::Lucene90(reader) => reader.get_block_start_pointer(block_id),
         }
     }
 
-    fn get_block_length(&mut self, block_id: i64) -> Result<i64> {
+    fn get_block_length(&mut self, block_id: i64) -> Result<usize> {
         match self {
             FieldsIndexEnum::Lucene90(reader) => reader.get_block_length(block_id),
         }
     }
 
-    fn get_start_pointer(&mut self, doc_id: i32) -> Result<i64> {
+    fn get_start_pointer(&mut self, doc_id: i32) -> Result<usize> {
         match self {
             FieldsIndexEnum::Lucene90(reader) => reader.get_start_pointer(doc_id),
         }
